Refactor duplicate process internally in duplicator service
